requirements.txt
# Core dependencies for POFJSP algorithm
numpy>=1.19.0
scikit-learn>=0.24.0

# Data handling and storage
pyarrow>=8.0.0
pandas>=1.2.0

# Testing framework
pytest>=6.0.0
pytest-cov>=2.10.0

# Configuration management
hydra-core>=1.3.2
omegaconf>=2.3.0

# Optional: For future visualization and data handling
matplotlib>=3.3.0
seaborn>=0.11.0

# Optional: For advanced optimization and analysis
scipy>=1.6.0

# Development tools (optional)
black>=21.0.0
flake8>=3.8.0
mypy>=0.800
